Setup ToDo object class
This class allows to create todo objects containing a title, description,
due date and priority.
It is possible to change the "done" state of the todo and each time the
done state is changed, the class checks if it was done in time by
comparing the current date with the todo's de date.